roles/sshkey: add role
This role can be used to either copy an existing SSH private key to a
target or to create a new keypair in the target and then, optionally,
upload it to GitHub.

Uploading to GitHub requires that the host (not the target) have the
github CLI installed (gh) and already authenticated.